Update test case service class.
