+++ dev/src/TestManagement.API/TestManagement.API/Features/TestCases/Get/GetTestCaseResponse.cs
@@ -1,38 +1,26 @@
-﻿namespace TestManagement.API.Features.TestCases.Get
+namespace TestManagement.API.Features.TestCases.Get
 {
     /// <summary>
-    /// Response DTO returned when retrieving test case information (latest version per case).
+    /// Response DTO that contains a TestCase and its single latest version.
     /// </summary>
     public class GetTestCaseResponse
     {
-        /// <summary>
-        /// Identifier of the test case version record.
-        /// </summary>
         public long Id { get; set; } = 0;
 
-        /// <summary>
-        /// Business code of the test case.
-        /// </summary>
         public string Code { get; set; } = null!;
 
-        /// <summary>
-        /// Name/title of the test case version.
-        /// </summary>
-        public string Name { get; set; } = null!;
+        public ICollection<TestCaseVersionItem> Versions { get; set; } = new List<TestCaseVersionItem>();
 
-        /// <summary>
-        /// Description of the test case version.
-        /// </summary>
-        public string Description { get; set; } = null!;
-
-        /// <summary>
-        /// Identifier of the test level associated with this version.
-        /// </summary>
-        public long TestLevelId { get; set; } = 0;
-
-        /// <summary>
-        /// The version number for this test case version.
-        /// </summary>
-        public int VersionNumber { get; set; } = 0;
+        public class TestCaseVersionItem
+        {
+            public long Id { get; set; } = 0;
+            public string Name { get; set; } = string.Empty;
+            public string Description { get; set; } = string.Empty;
+            public int VersionNumber { get; set; } = 0;
+            public long TestLevelId { get; set; } = 0;
+            public bool IsLatest { get; set; } = true;
+            public System.DateTime CreatedAt { get; set; }
+            public System.DateTime UpdatedAt { get; set; }
+        }
     }
 }
